collector/pic32: separate special collector
